Move starlark data structures to crate::utils::starlark
